void logic in VoidMetadataEmitter
